requirements.txt
pytube==15.0.0
python-dotenv==1.0.0
openai==0.27.8